requirements.txt
django==1.9.3

djangorestframework==3.4.6